Move `RecordBinaryHeap` tests into more logical location
