+++ crates/subspace-networking/src/utils/record_binary_heap.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+mod tests;
+
 use libp2p::kad::kbucket::Distance;
 pub use libp2p::kad::record::Key;
 pub use libp2p::PeerId;
